requirements.txt
discord.py
requests-html
